refactor: split ml_analysis.rs into 14 modular files
- Split 4654-line monolithic file into logical modules
- Created ml_analysis/ directory with:
  - architecture.rs (84 lines) - Model architecture analysis
  - memory.rs (200 lines) - Memory usage analysis
  - learning_rate.rs (247 lines) - Learning rate tracking
  - convergence.rs (1166 lines) - Training convergence patterns
  - gradient.rs (695 lines) - Gradient flow analysis
  - attention.rs (435 lines) - Attention mechanism analysis
  - ensemble.rs (463 lines) - Ensemble model analysis
  - quantization.rs (871 lines) - Quantization analysis
  - batch_norm.rs (112 lines) - Batch normalization
  - regularization.rs (107 lines) - Regularization impact
  - activation.rs (108 lines) - Activation patterns
  - weight.rs (101 lines) - Weight distribution
  - complexity.rs (403 lines) - Model complexity
  - mod.rs (29 lines) - Module exports
- Preserved all public API - no breaking changes
- Build verified successful
