+++ diffai-core/src/ml_analysis/mod.rs
@@ -1,0 +1,29 @@
+// Module declarations
+pub mod architecture;
+pub mod memory;
+pub mod learning_rate;
+pub mod convergence;
+pub mod gradient;
+pub mod attention;
+pub mod ensemble;
+pub mod quantization;
+pub mod batch_norm;
+pub mod regularization;
+pub mod activation;
+pub mod weight;
+pub mod complexity;
+
+// Re-export public analysis functions
+pub use architecture::analyze_model_architecture_changes;
+pub use memory::analyze_memory_usage_changes;
+pub use learning_rate::analyze_learning_rate_changes;
+pub use convergence::analyze_convergence_patterns;
+pub use gradient::analyze_gradient_patterns;
+pub use attention::analyze_attention_patterns;
+pub use ensemble::analyze_ensemble_patterns;
+pub use quantization::analyze_quantization_patterns;
+pub use batch_norm::analyze_batch_normalization_analysis;
+pub use regularization::analyze_regularization_impact;
+pub use activation::analyze_activation_pattern_analysis;
+pub use weight::analyze_weight_distribution_analysis;
+pub use complexity::analyze_model_complexity_assessment;
